requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
celery==5.3.4
openai==1.3.0
jieba==0.42.1
schedule==1.2.0
python-telegram-bot==20.7
pymongo==4.6.0
motor==3.3.2
aiofiles==23.2.0
aiohttp==3.9.1
httpx==0.25.2
lxml==4.9.3
feedparser==6.0.10
python-multipart==0.0.6
jinja2==3.1.2
websockets==12.0
bson==0.5.10 